+++ apps/scoutgame/app/(general)/info/builders/page.tsx
@@ -1,0 +1,11 @@
+import type { Metadata } from 'next';
+
+import { BuildersPage } from 'components/info/pages/BuildersPage';
+
+export const metadata: Metadata = {
+  title: 'Builders'
+};
+
+export default async function Builders() {
+  return <BuildersPage />;
+}
