+++ apps/scoutgame/app/(general)/info/partner-rewards/celo/page.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from 'next';
 
-import { CeloPage } from 'components/info/PartnerRewards/CeloPage';
+import { CeloPage } from 'components/info/pages/PartnerRewards/CeloPage';
 
 export const metadata: Metadata = {
   title: 'Celo Partner Rewards'
