+++ apps/scoutgame/app/(general)/info/partner-rewards/moxie/page.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from 'next';
 
-import { MoxiePage } from 'components/info/PartnerRewards/MoxiePage';
+import { MoxiePage } from 'components/info/pages/PartnerRewards/MoxiePage';
 
 export const metadata: Metadata = {
   title: 'Moxie Partner Rewards'
