+++ apps/scoutgame/app/(general)/info/partner-rewards/page.tsx
@@ -1,9 +1,11 @@
 import type { Metadata } from 'next';
 
+import { PartnerRewardsPage } from 'components/info/pages/PartnerRewardsPage';
+
 export const metadata: Metadata = {
   title: 'Partner Rewards'
 };
 
 export default async function PartnerRewards() {
-  return null;
+  return <PartnerRewardsPage />;
 }
