+++ apps/scoutgame/app/(general)/info/scouts/page.tsx
@@ -1,0 +1,11 @@
+import type { Metadata } from 'next';
+
+import { ScoutsPage } from 'components/info/pages/ScoutsPage';
+
+export const metadata: Metadata = {
+  title: 'Scouts'
+};
+
+export default async function Scouts() {
+  return <ScoutsPage />;
+}
